Refactor : "SMS 전송 로직 개선" - 중복 코드 개선
Related to: #15
