+++ build.gradle
@@ -21,7 +21,7 @@ repositories {
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '5:5.1.3'
+    implementation group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.13'
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
